+++ ClickHouse.Driver.Tests/AssemblyInfo.cs
@@ -1,4 +1,4 @@
 ﻿// Needed due to issues with floating point comparisons arising on ARM architectures
 using System.Reflection;
-[assembly:DefaultFloatingPointTolerance(1e-16)]
+[assembly:DefaultFloatingPointTolerance(1e-15)]
 [assembly:AssemblyKeyFile("../sgKey.snk")]