+++ ClickHouse.Driver/Types/LineStringType.cs
@@ -1,0 +1,11 @@
+namespace ClickHouse.Driver.Types;
+
+internal class LineStringType : ArrayType
+{
+    public LineStringType()
+    {
+        UnderlyingType = new PointType();
+    }
+
+    public override string ToString() => "LineString";
+}