+++ ClickHouse.Driver/Types/MultiLineStringType.cs
@@ -1,0 +1,11 @@
+namespace ClickHouse.Driver.Types;
+
+internal class MultiLineStringType : ArrayType
+{
+    public MultiLineStringType()
+    {
+        UnderlyingType = new LineStringType();
+    }
+
+    public override string ToString() => "MultiLineString";
+}